New project on Antigravity/nova_career_pilot/requirements.txt
streamlit
boto3
PyPDF2
python-dotenv
pandas
plotly
